Updated mainwindow.ui
This update adds actions like Settings, Themes, Light, Dark, Salmon and Hacker to the .ui file. This update also adds implementation for more customization in the near future, my goal is for people to be able to customize using a .qss file, config, during building of the app or UI navigation.